move provider into its own module
