+++ docs/source/api/root.rst
@@ -10,4 +10,5 @@ ci_exec
    core
    colorize
    patch
+   provider
    utils
